Bug. Fix: Avoid SIGSEGV when sending a status-only request to a LocalPeer
Because of the behavior change where we no longer sleep in the reactor thread on comm.
failures, RaftConsensusTest::WaitForCommitIfNotAlreadyPresent() is now sending actual
status-only requests. This brought to light a bug in the Peer implementation where
we would reset the heartbeater without checking that one was set first (the local version
has no heartbeater), leading to a SIGSEGV.

This fixes this by adding a check that the heartbeater is set first and
makes the LocalPeer test in consensus_peers-test.cc send a status-only message.
This test fails without this patch and no longer fails with it.

Change-Id: Iae8fe9899e308d8f793f46119c5a24b32c08e7fd
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2919
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>