mvcc: fix tracking of snapshots when transactions start out-of-order
This fixes an issue whereby we were incorrectly adjusting the state of
MvccSnapshots when timestamps were assigned out-of-order (a consequence of
commit-wait transactions).

The new design tracks a lower bound timestamp below which all transactions
are committed. A snapshot maintains a set of any transactions above this
timestamp which already committed, rather than tracking _uncommitted_
transactions. The MvccManager, in addition, tracks the in-flight transactions,
so it knows when it can move up the all_committed_ lower bound.

Because this commit added a new call to clock_->Now() in the commit path,
several other unit tests had to be updated. Those unit tests rely on the logical
clock, so the extra Now() calls increased the timestamps expected by the tests.

Change-Id: I6e77ee35a8c76a82123fed8e58804843f2ec5522
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2927
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>