client: new API for retrieving scanner rows
No one liked the uint8_t pointer approach, so here's a new one. Goals:
1. Expose a nicer JDBC-like API for retrieving cell values.
2. Minimize amount of memory allocation and copying.
3. Don't depend on ContiguousRow and friends, which are deeply intertwined
   in the backend code.

To that end, here's KuduRowResult (prefixed with Kudu because it's
user-facing). It copies large swaths of PartialRow and some parts of
ContiguousRow and ContiguousRowHelper.

To avoid allocating each KuduRowResult individually, I had to use an Init
function in lieu of an argful constructor. This looks weird; if there's a
better way to handle this I'd love to hear it.

Change-Id: I1a4395e4f16bddc8a5912da9c1cb4a0f3026c29c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2928
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>