KUDU-337 - Fix flaky test in raft_consensus-itest
TestFailedTransaction is currently flaky. This happens because there
is nothing in the tablet (the transaction failed) so the AssertAllReplicasAgree()
call doesn't actually test that all transactions have executed.

The final solution to this and other quiescing problems is to implement proper
transaction abort (filed KUDU-341 for that). In the meanwhile, this makes it
so the test isn't flaky anymore by adding an actual successfull write after the
failed one and testing that all replicas agree that one went through.

I could reproduce the error before but can't now.

Change-Id: I1384c761ad7f48a2355d00b16c68d5b0a7d85a6f
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2950
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins