linked_list-test: loop test even if the linked list verification fails
Since the test writes to multiple tablets, and we are scanning without
any consistency, it's possible that on the first scan we may see
a broken linked-list chain. This would cause a non-OK Status response.
We need to loop and retry in this case, not just in a case where we
have too few rows.

Change-Id: Id6cd7e4bfc303df8eb5b97cdabe4a4fc7416e8ef
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2957
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins