KUDU-264. Re-implemented some methods in line_item_dao to use client API
Since the C++ API uses already-created PartialRows, it is not possible for
the user of the LineItemDAO classes to just pass in whatever row they
feel like and expect the class to handle it.

The smallest API change to this is to pass a function pointer void(PartialRow*)
instead, where the visitor pattern allows the user of the classes to just tell
what changes to the PartialRow are desired. This has a beneficial side
effect of actually increasing safety (only the associated schema will be
used) and ensuring the lifetime of the PartialRow used.

The API change required some minor changes to some of the other classes as
well.

Change-Id: Ia9c687334e07c40f81b4c3fea9d2b22e9ab6f413
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2988
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins