Consensus: Make the queue always accept commits
Currently the queue refuses all messages when full, but it should accept
commits since they are required for transactions to complete and won't
grow the queue out-of-bounds without the addition of more replicate messages.

This patch fixes this by allowing commits to always be accepted even when the
queue is full. Moreover this also adds a NoOpResponsePB to help in testing this
and renames NoOp to NoOpRequestPB.

Change-Id: I309ab973c99b68f86121044361f79096a6efb2e9
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2979
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins