consensus_queue: always send at least one op
This fixes the possibility of a consensus stall in the case that a large
op makes it onto the queue. Previously, if the op were larger than
the configured max batch size, it would "stick" there and never get
sent. Thus, the transaction would stall, never commit, and things
would cascade from there.

Now, we always send at least one op if there are ops available.

Change-Id: Id174bcec9538119c1ad34e74debec255f9a5bd34
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3011
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins