[java client] Harden RowResult#getString()
There was a TODO in that class about figuring the int/long mess. The problem
is that the c++ side encodes a Slice in two longs, but we only need ints.
Tried changing this but gave up, so ended up only making sure that the Java
client will only see good ints.

Change-Id: I2d2516c0bee24cfc634d92417a43449ce5944111
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3008
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>