Disabled unit test with multiple row mutations for the same key in a batch.
Currently, there is no order imposed within a batch by the client code.
Before attempting a fix, a contract needs to be decided upon for the
strength of the ordering that will be supported within a batch.

There are tradeoffs to consider - we may not wish to force clients desiring
ordered operations to flush, since this may be desired behavior but
with flushing will be expensive.

Change-Id: I4b0a25f69a751eaf5aa333ab409621ea8fb88178
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3055
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>