[java client] Finalize KuduClient#shutdown()
There was a TODO in that method about flushing the buffers, and this patch does
that. First, KuduClient needed to be able to access those buffers in KuduSessions
so we're now keeping track of them in a HashSet. Then the concept of a "closed"
client was added so that it doesn't get new sessions while trying to close down.
Finally, code was added to close all the sessions and return a Deferred to keep
track of all the operations that need to be flushed.

Change-Id: I14f4b30a6d5583129407c28b56f133eed88a40cd
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3041
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins