Move log path knowledge to FsManager
We need the log file / directory path information in FsManager in order
to use it from catch-up / bootstrap.

Also, name log files as wal- instead of log- in order to make it more
obvious to users that they are not logs that may be compressed or
archived / deleted. This prefix is also part of the FsManager class.

Change-Id: I81aca34fa7d38c15c15e4c252636e6be38bf6622
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3045
Tested-by: jenkins
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>