Make mvcc allow for point-in-time snapshots and "offline" commits
This adds the ability for Mvcc to allow for point-in-time snapshots and "offline" commits
i.e. snapshots that are created at a point-in-time (likely the past vs. reading the
current clock value) and are committed without advancing the 'all_committed_before_'
watermark. This is useful for replicas executing transactions with the same timestamps
as the leader and for bootstrap, where we want to exeucte transactions whose timestamps
are in the past. In order to clear state and advance the watermark clients must
call explicitely 'OfflineAdjustCurSnap(Timestamp)'.

The behavior when not using the new methods remains the same.

Change-Id: Ided2abb8f84bc60f51fe9141f81e17e80b26c259
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3123
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>