LOG_TIMING macros should log even with an early return
Since the LOG_TIMING macro does not use RAII, if there is an early
return there is no timing message. Change the action to happen from the
destructor (if it hasn't already happened before).

Also, add SCOPED_ versions of the timing macros, so allow for timing the
rest of the code in a block.

Also, instead of printing the message "Times for", which is kinda weird,
print "Time spent" as the prefix instead.

Change-Id: I11015d257d444df44e5aa5aa5c1cc02aa955af1e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3084
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>