KUDU-114 Part I: Memory accounting by subsystem
Ports MemTracker from Impala with several changes to work
with Kudu and fit the Kudu use case more closely.

NOTE: code style in this revision is somewhat different
from the rest of the Kudu code base, as when possible code
was cut-and-pasted from Impala with as few changes as
possible.

TODO (Part II): use this for several core classes,
                integrate with /memz.

Change-Id: If32800eb8f3b48009ad16c1e8f6090c73acdb97a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3118
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>