KUDU-217: mechanism for scheduling tasks to run after a delay
This commit adds support for scheduling arbitrary tasks to be run on the
reactor thread after some fixed delay. I played with a number of designs and
eventually settled on this one because:
1. It reuses the most code, and
2. It avoids duplicating code (e.g. telling the reactor thread to do
   something, handling shutdown, etc.).

We discussed whether this should be implemented in the reactor or not. Even
though a standalone scheduler could be more concise and be reusable by more
components, in the end I went with the reactor so that I could reuse a
bunch of existing code (libev loop in each reactor, ReactorTask, etc.). This
is fine for the C++ client which already has a Messenger, though it may be
problematic for future use-cases.

Change-Id: Id0c38e42127fad18e384410677d1ab2bd8438223
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3107
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins