Improve stringification of WriteTransaction
This fixes an issue where invalid timestamps would be stringified in
a transaction as if they were a real timestamp, which led to some
confusing output. Now, a DCHECK ensures that we don't read the timestamp
of a transaction until it has been set, and when stringifying, we
write '<unassigned>' instead of showing an invalid timestamp.

This also changes the timestamp to be an optional field in a commit
message, since aborted transactions do not actually have a timestamp
set yet.

Change-Id: Ifcfcbc96f57c62c004a97a0c091099e3e191a943
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3112
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>