KUDU-357: add libunwind to thirdparty
Todd found that if the libunwind dev package was installed at the time that
glog was built, the static build of kudu breaks due to unresolved libunwind
symbols. The recent shlibs patch surfaced this, though it's not clear why;
I saw the issue with a repo checked out to before that patch too.

In any case, we have several options:
1. Disable libunwind support in glog.
2. Tell kudu developers not to install libunwind.
3. Force them to install it and make it a dependency of glog.
4. Add it to thirdparty as a first class citizen.

I've gone with option #4 because it's safest and gives us the most
flexibility when it comes time to release (e.g. shared vs. static libunwind).
It's also the easiest for everyone else.

Change-Id: I6ba917716824f8dc86631cdaad024ce2837ceb56
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3116
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins