Caught potential bug in client test.
The randomized row operations testing in client test (TestRandomWriteOperation)
made sure never to send a batch with a delete/insert sequence.

However, this still allowed the possibility of an insertion/mutation sequence,
the order of which may be inverted (see KUDU-352). We were not seeing these
errors here before since the probability of the same row being randomly
selected in the same batch (of 50, with 1000 rows) is on the order of
1/100. Furthermore, as the time between the applications of the insert
and the mutations gets larger, the batcher reactor threads (whose
order inversion causes the issue in KUDU-352) have an increasing chance
of finishing in their proper sequence.

But just in case, the batch is now restricted to 50 unique rows.

Change-Id: Ibed0981b189e516830cdb49b875e60e44551dd1a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3121
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins