build: fix linkage against libkudu_test_main.so
As test_main.cc calls timer_ functions from librt, it ought to be explicitly
linked against librt. The weird thing is that my system and Jenkins cope
with the lack of explicit linkage just fine. The only observed failure was in
Ubuntu Saucy, and even then not for every unit test.

This also removes an errant set(LIBS ...) line that has no effect anymore.

Change-Id: I95082544ca2b4acf091af043cfe0c857c1bd20ee
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3129
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins