[java client] Define an interface for sync and async clients
This is the first step towards having a synchronous client. The design is to
have both KuduSession (maybe rename it) and SynchronousKuduSession that present
a very similar API. Both will have apply(), but the signature will be different
since a sync API will have blocking methods that throw exceptions if something
goes wrong, wheras everything is passed through the callbacks for the async one.

In this patch, we extract all the common configurations for a session into a
common interface.

Change-Id: I6fb2c50d2c60b1484e9eac5f12137bf9c3b5da82
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3167
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>