KUDU-131 - Part 3 - Make replicas and bootstrap reuse the timestamps of the original transactions
With this patch bootstrap now uses the timestamps found on the commit messages to
replay transactions, i.e. the result of a snapshot scan at a timestamp before and
after bootstrap is now the same.

This also adds similar support to replica transactions, making them use the timestamp
of the original leader transaction, i.e. the result of a snaphot scan at a replica
and at the leader is now the same.

This adds testing for this by making linked_list-test select a sample of
snapshot during the creation of the linked list and then making sure that each
snapshot is the same across replicas and reboots.

Advancing the safe timestamps is actually a tricky proposition, so I'll include that in
a following patch (which might make this one a performance regression).

Change-Id: I8bf5cb9347eba6b29e9ec485dc2c1c7ca6df6d9a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3188
Tested-by: jenkins
Reviewed-by: Vladimir Feinberg <vladimir.feinberg@cloudera.com>