KUDU-131 - Reuse timestamps when replaying transactions
This makes it so that we reuse timestamps when we replay transaction on replicas
or on bootstrap.

Timestamps are obtained on the prepare phase, on the leader side, and embedded in
ReplicateMsgs when they are submitted for consensus replication.

When replicas receive transactions from the leader, instead of assigning those
transactions timestamps obtained from their own clock, they instead assign the
timestamp assigned by the leader.

An identical mechanism is used, on bootstrap, to make sure that transactions
retain the timestamps across reboots, thus making sure that snapshots scans
at a timestamp always return the same data.

This also makes linked_list-test collect timestamps when doing writes and then
perform snapshot scans at each of the writes to make sure all replicas match.

TODO: Handle safe time advancement properly. Right now replicas don't advance
the safe timestamp at all, the only safety mechanism that is in place is the
fact that replicas update their own clocks with the timestamp of the latest
message received by the leader, thus making sure that if a replica becomes
leader it won't assign timestamp that is lower than any that of any message
received by the leader.

Change-Id: Idefbd085d3154010d98cfadac0af8d2f7697f833
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5939
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins <cloudera-hudson@gerrit.cloudera.org>