KUDU-177. fix handling of major delta compaction and bootstrap
This substantially restructures how major delta compaction is wired into
metadata and rowsets. The old code generated a new DiskRowSet which it
swapped in at the table level. However, this broke assumptions with
bootstrap, since the commit messages ended up with the wrong rowset
IDs, etc, and edits would not be properly replayed.

The new code performs the major compaction within the DiskRowSet and
then re-opens the base data and swaps in the new DeltaFileReaders
upon completion.

This also fixed a couple of other apparent bugs along the way which are
covered by the new unit tests.

It's likely that there are still bugs as this relates to "alter table",
especially in concurrency scenarios. That is to be addressed in a follow-on.

Change-Id: Ic1a41b61b20df7d1c2eab07f8c9f67393ac52937
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3182
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>