pstack unit test should restore stdout
/dev/stdout does not exist after stdout is closed, and we were not
checking the return code from freopen(), which was indicating an error
with errno containing ENOENT. Therefore, the TestPstackWatcherRunning
test was apparently borking stdout.

Print unit test pstack output to stdout (for posterity, and to ensure
that printing to stdout works).

Also add some method comments and a minor formatting tweak.

Change-Id: Ib0622454cf731ffe509815db0c5b2d58e1031558
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3192
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>