[java client] Synchronous client as a simple wrapper around KuduSession
This patch adds a new class, SynchronousKuduSession, which basically just adds
joining on top of KuduSession's methods. It also adds the handling of
PleaseThrottleException by just re-applying the Operation.

This patch also shows that we need to post-process the exceptions in KuduSession
instead of sending an opaque Exception and then have the user play guessing games.

We'll also need to bring KuduScanner into KuduSession instead of KuduClient.

Change-Id: I993bdf0869d9952a333b53dd94d34143fbf636a6
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3204
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins