KUDU-46: Deboostify ColumnRangePredicate
To ship a safe C++ client, we need to ensure that there are no references to
boost in public, user-facing headers. Boost is notorious for breaking
backwards compatibility and the presence of Boost in arguments, return
types, or member variables raises the chance that the user's application
will break in bizarre ways if they link against a different version of boost
than our own.

The ColumnRangePredicate case is a shoe-in; its boost::optional fields are
pointers and thus can handle "optionality" with the NULL value. Granted,
it's not the best API and should change in the (near) future, but even then,
boost::optional won't be returning.

Change-Id: I83c9e902f3f407ad625e7a2cab7b8b92a8cb9b72
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3207
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>