metadata: remove delta IDs for delta files
This somewhat-messy commit removes the usage of delta "ids" to refer to flushed
delta files. This concept was somewhat broken when we introduced delta
compaction, because the "ids" were no longer unique identifiers. For example,
consider the following sequence:

1) Flush DMS: delta file has index 0
2) Flush DMS: delta file has index 1
3) Flush DMS: delta file has index 2
4) Compact delta file 0 and 1

The old code didn't really know what ID to assign the delta file created at
step 4, and we ended up with metadata updates using indexes in the delta file
list to refer to the files instead of the IDs.

To cut a long story short, the IDs were not in fact useful, when we already
have block IDs to uniquely identify files. So, DeltaFileReaders now store their
block ID, and the metadata update uses the sequence of compacted block IDs to
specify what to replace after a compaction.

Some further cleanup can be done after this point (eg removing the redundant
"path" parameter from DeltaFileReader) but I wanted to minimize the size of
this commit and keep it to a fairly straight "translation" of the int64 IDs to
block IDs.

Change-Id: I0a034f2e5d747776c5caa8352ac4fd5bdd51eb3b
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3237
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>