KUDU-341 - Allow to abort transactions
Transaction abort is implemented the following way:
- On the LEADER side only the status is set. When the next task completes
  reading a non-OK status will cause the transaction to abort. This approach
  allows to add abort without actually cancelling running tasks. When the
  next task completes, if the transaction went past the point of no return
  it will try to replicate an actual OP_ABORT message.
- On replicas abort is only called on consensus shutdown and does nothing
  but delete the pending transactions (after making sure that no more transactions
  can be started).

This takes care of the test failures where there was a message like:
"waiting for X outstanding transactions", such as the failure in
DistConsensusTest.TestInsertOnNonLeader. As it turned out the hanging transactions
were replica transactions. Under heavy load this problem was moderately reproducible
after the fix it no longer is (ran about 500 * 10 parallel iterations without
issue).

There is another issue apparently similar but unrelated where the leader
keeps trying to update a replica but replica replies with ServiceUnavailable
and there are no "Waiting for X oustanding transaction" messages. This issue
is not solved by this fix.

Change-Id: I46f379b595c9eea7ecfd13e27931fb92004ba883
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3233
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>