KUDU-261: Add SyncDir() to Env, use in Log
Fixes KUDU-261: adds a method to fsync() a directory and syncs
the log directory after a newly allocated segment is renamed.

Note: this was committed earlier and reverted: the earlier
version expected return value of 0 from open() to mean normal
behavior (which was not correct, as open() returns the fd
number).

Change-Id: Ie5ed4fc940a650be71745167867c1a00ac622066
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3235
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>