benchmarks: fix bug where git hashes would be evaluated relative to each other
This fixes a bug where, if you ran 'benchmarks.sh -local HEAD~3 HEAD' it would
actually run the benchmarks twice on HEAD~3, since it checked out each ref in
turn as specified on the command line.

This fixes it by running rev-parse on each of the refs up front before checking
out any of them.

Also added the ability to pass NUM_SAMPLES as an environment variable on the
command line to make it easier to do a quick benchmark diff locally.

Change-Id: Id287d81c4b8ddfa689506ad604ddb6ace1dfac91
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3242
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>