KUDU-365: retry RPCs that failed due to server backpressure
Service pool logging revealed pathological batching behavior. Consider the
following simple scenario:
1. A fresh client is opened, with a fresh session.
2. 1000 writes are applied to the session.
3. The session is flushed.

Because the client is fresh, the first few writes will issue master lookups.
These lookups can take time, and if we make it all the way to the flush
before they finish, the writes will be flushed immediately once the lookups
are done. In this case, "immediately" means, "as soon as we have a tablet
location", meaning, without any batching.

Rather than fixing this case, let's add generic retry behavior, which will
be useful in other contexts too. If we get a SERVER_TOO_BUSY during an
operation, we schedule the RPC to resend in a few ms (more for each
subsequent attempt). We stop retrying not by the number of failures, but by
exceeding the original write timeout.

The generic retry behavior is also applicable to master lookups. You might
think that the master lookup permitting behavior would obviate the need for
retrying, but:
1. There could be multiple clients performing lookups simultaneously, or
2. The master responds to other RPCs (like TSHeartbeat), which can occupy
   space in the service pool.

I implemented the same thing for master lookups, and it's very clear from
the diff that we need a common RPC layer to handle this; that's next on my
agenda.

Change-Id: I17ea659550efc118a1ffdb33e3e37d55121e9a50
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3263
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>