monodelta: forbid uninitialized deltas
It's not currently possibly to distinguish between a "real" MonoDelta (e.g.
build with FromSeconds) and an "unitialized" one (built with the default
constructor). There are cases, though, where it's useful to tell them apart.
So let's make that possible, using kint64min as the reserved "magic" value.

What are uninitialized deltas good for? Not much. They're basically
placeholders, waiting to be reset with the assignment operator.

Change-Id: I69cb2c0bbd7e893cd2e6ac5ccd58bbfe176bcedf
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3260
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>