[java client] Get rid of all the warnings
This patch changes how shutdown works, actually it works almost the same
way except that we're not lying about what it does.

TabletClient#decode is still not typed but has note on why. It now does a
better job at verifying the inputs.

KuduRpc#deserialize now returns a pair of type R and Object, the latter
being the error PB the server can send.

Change-Id: Icdb7970d398a9f3667018c237611f67044f27a9a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3264
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>