KUDU-365: only flush when last outstanding lookup finishes
It's possible to flush a session while a large number of writes are still
looking up their tablet locations. Without special handling, each write
will be flushed as its lookup completes, resulting in a steady stream of
small RPCs to the TS. Given enough clients, these streams can form a
"thundering herd" of RPCs that can overwhelm the TSes.

To address this, let's keep track of the number of outstanding lookups, and
flush only when the very last one is finished. The tricky bit: we now need
to flush all per-TS buffers together, because the last lookup will only
belong to one per-TS buffer.

Change-Id: I55fa9935ecadab5f6d6d01efee3bef4c84d29cda
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3269
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>