Add an RPC-based insert stress test against TS
This adds tablet_server-stress-test, which spawns a configurable
number of threads which make Insert() RPCs against a tablet server.
Upon completion, it reports some measures of throughput and a
latency histogram.

This is useful both as a concurrency stress test of the whole TS stack
as well as a simple benchmark.

Change-Id: I0438dccdbf0875098ffdba5e49363e2dcb541506
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3276
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins