KUDU-116: Server-wide consensus_queue memory usage
Use MemTracker to keep track of and limit server-wide and
per-tablet consensus queue memory usage. This information is
also available via /memtrackerz.

New flags are added to allow specifying server-wide size limits.

If TcMalloc is enabled, then whenever consensus queue is trimmed the
released memory is returned to the OS.

Change-Id: I20605ac36fd795249446d1353c94d4bb143f98bb
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3285
Tested-by: jenkins
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>