KUDU376 - Fix deadlock in transaction abort
Transaction abort was deadlocking because we were holding the consensus
lock while aborting at the same time as transactions could be holding the
transaction state lock while committing (and thus needing the consensus lock).

This solves that by restraining RaftConsensus::CancelPendingTransactions()
to only cancel those which are not in-flight, i.e. whose Apply has not
yet begun and thus have no risk of trying to attempt a simultaneous commit.

Ran a bunch of parallel runs and the problem disappeared, as expected.

Change-Id: Icc2b25cadd7ed21fe7af2d844677c8ed14fae936
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3290
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins