Add a Status::Incomplete()
Usually we either block indefinitely until a task is complete, or
wait a certain period and get a Status::TimeOut() in the end if the task
isn't completed yet. In certain cases though we also need to check on the
status of the task without actually blocking. This new Incomplete Status
can be used in those cases.

Change-Id: I77765a8ea5b08b17f3158b69ccd5a79fcb86c358
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3339
Tested-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>