KUDU-375 - Make the AlterSchemaTransactionState set the AlterSchemaRequestPB* to NULL on commit
This was caused by a trivial race where the request would be deleted by the RPC layer
while the transaction was still alive. Since the transaction state held a pointer to
the now deleted request we would SIGSEGV.

This patch fixes this by explicitely setting the request pointer to NULL on commit, meaning
the ToString() method in AlterSchemaTransactionState will no longer try to print it and
the segmentation fault we would see in KUDU-375 no longer happens.

This also includes similar changes for all transactions types (for both 'request_' and 'response_').

Change-Id: Ie3440c71b93afb290a94cac3de5d9ec78ba56cbf
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3342
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: David Alves <david.alves@cloudera.com>