Removed epsilon from compaction selection algorithm files.
The only places where the cdf value epsilon from the RowSetInfo class was used
were the svg_dump.cc and compaction_policy.cc files.

In the svg dump, it was used through the Intersects() method to make sure
that graphics did not overlap. Thus, the epsilon was not really necessary.

For the compaction_policy.cc, the epsilon was used in an unrelated context to
the addition of the cdf widths in different orders, which has a different
epsilon bound than that from generating the cdf min/max keys. A "magic"
epsilon value was put in (rigorously derived epsilon was not worth the trouble
for a DCHECK).

Change-Id: I4ef4be6d8b4ff2993f9936339415a1444b26aa69
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3345
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>