[java client] Move the SERVICE_UNAVAILABLE handling down to TabletClient
Lost in the changes done to returned types, we missed the bit where Batch's
callback was still trying to retry itself if it was getting SERVICE_UNAVAILABLE.
Since TabletClient now redirects the Errors to the errback, that block of code
was dead.

This also has the effect of retrying anything that gets SERVICE_UNAVAILABLE,
it's a bit coarse but I don't think it will do anything bad.

Change-Id: I8b1735720a3382adc65c616a320aaccca70bd173
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3347
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>