KUDU-46: use friendship to reduce client public dependencies
Scattered around the C++ client are a number of references to classes that
are only used internally. Let's make these references private and use
friendship to expose them to internal classes as needed.

Once the client is fully PIMPLed, all user-facing header files will be
bereft of these private members.

Change-Id: Ib9a66dee900d767b1f30ff5cc203537d8b12e5b0
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3350
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins