Fix various corner-case bugs in RowSetTree slice comparator.
When there is only one key in a rowset, there was a chance that
the START endpoint would be placed before the STOP endpoint. This would
be caught only in the traversal by rowset_info.cc rather than during
construction. Changing the endpoint comparator to use a lexicographical
ordering (Slice -> Rowset -> Endpointtype) will avoid the issue.

Another corner case with one key is that the shared prefix between
the min and max keys of the RowSet keyspace interval is the entire key.
This would fail an assertion by the Slice's operator[].

Change-Id: Idf1e38595bae0fb36e42b068f550d8889dca7ccd
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3356
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins