[java client] Better return types, part 4/5, fixing the locateTablet mess
This is the last part of the code that Deferred<Object> because we were trying
to handle both callbacks and errbacks in the same Callback. This used to work
"ok" since everything is an Object, but with more concrete types we can't use
this trick anymore.

This patch adds some more error handling, clears some erroneous ones, and
notably removes the handleLookupExceptions method. With the recent changes
in TabletClient, it's unclear if that method was even used. This means that
errors that happen during tablet lookups or IsCreateTableDone calls will be
directly thrown back to the user. Two TODOs were unfortunately added for this.

Change-Id: Icd19f9aff39cdf1b483760efebfaa79a2c86f838
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3362
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>