[java client] Better return types, part 5/5, PBs be gone
This final patch removes the sending of PBs from all the public-facing APIs.
They are replaced by children of KuduRpcResponse which makes it easy to provide
information such as the elapsed time for all RPCs.

This patch also removes the PingRequest, it was mostly a test used when the
client was being written in the beginning and now its becoming a burden.

This patch also changes the way we parse the Messages, we now skip building
them completely.

Change-Id: Idbb03270f70bde05b06a6231de9b2fe7cc16f4cd
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3367
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>