Make TransactionDriver::ToString() acquire the lock
A previous patch set the request_ in AlterSchemaTransaction to NULL on commit()
which reduced the chance of getting a SIGSEGV on TransactionDriver::ToString()
but apparently not to 0. This because there was a race to address between
the committer thread and the printing thread.

This patch addressed this race by making ToString() acquire the lock
and adding a ToStringUnlocked() method that is used where relevant.

Change-Id: I483095b01f1049ec77377d2b85652ad03767e144
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3368
Tested-by: David Alves <david.alves@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>