write_transaction: remove multiple constructors
The three constructors were basically the same, and we generally try to avoid
overloading. This removes the multiple ctors and replaces them with a single
one.

This uses default arguments in the ctor, which isn't really a preferred style,
but I think it's better than three nearly-identical overloads.

Also moved the dtor to the .cc file and marked it virtual, since it's already
virtual by virtue of its superclass anyway.

Change-Id: I146100691c3e8e33d72cfe88f9c9c579cea1ecda
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3371
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins