Get rid of PeerMessage, make OperationStatusTracker take ownership of the OperationPB
This gets rid of PeerMessage, a class used in consensus queue to store both the operation
and the tracker, and instead makes the tracker take the operation in its ctor and makes
the queue store that directly. This simplifies things and ownership and allows us to
print the operation when something happens to the tracker, for debugging.

Change-Id: I8b0fdac644d22ca151f154c2bdd5edf4a2bc3792
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3413
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>