compaction: fix bug in knapsack solver which broke budgeted compaction
In cf3bea1189a37ac1086209d59880627c07a466e7, KnapsackBlackboard::Advance()
accidentally introduced a typecast from the knapsack "value" type to integer.
This didn't cause tests to fail, because the tests used an integer value type,
so it was equivalent. However, in real usage, the value_type is a double,
so this actually broke quite a bit of stuff.

We didn't catch it quickly because the compaction policy code lacked any
proper unit tests. This patch adds a new unit test for the compaction policy
based on a specific instance that I saw breaking (i.e not compacting) in a
real server situation.

This patch also changes the knapsack solver test to use double values, to more
closely track what we do in real life.

As a happy side effect of this change, it appears that the knapsack solver
benchmark runs marginally faster by avoiding the type casting back and forth
between int and double.

Change-Id: I58e676f26f8bb8865132454d544727745b3c02e1
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3415
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Vladimir Feinberg <vladimir.feinberg@cloudera.com>