KUDU-389: ensure initial inserts are flushed before further changes
In the WAL of a failed test run, the DELETE of key=239 came before the
INSERT for that key. Turns out it's because we never explicitly flushed the
initial inserts for the test. If the first change in the loop was a DELETE,
we could hit KUDU-352 and add the DELETE to the batch before the INSERT.
That's also why this was rare: the test had to 1) choose (at random) to
change an inserted row, and 2) get hit by KUDU-352.

Anyway, the fix is to reorder the "correctness checks" to come before any
changes. Why not an explicit flush for the inserts? Because we want to
verify the validity of the inserts (using the correctness checks) too.

Change-Id: I84bb3e0c57e0cfd830a986bd8099214e9e59a8f5
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3420
Reviewed-by: Vladimir Feinberg <vladimir.feinberg@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>