Fixed compilation bug in env-test.cc.
On gcc 4.8, Ubuntu 14.04, calling open() where the creation of a file may
result causes a compilation error since the file mode for user privleges was
not specified.

On other systems a default mode was provided, but this is not required
behavior. This commit adds a file mode (and changes the call to creat()
to avoid future confusion).

Change-Id: I1ec1da66bfd61b334fddb6b789ba65081cd286cb